Revert changes to custom device xml made in cb7a0537d760aa0f045108f802ba75ce145668e0
The custom device XML references pages which are provided by VeriStand,
which are needed at runtime. These VIs are not included with the custom
device, and so the pages were erroneously removed from the xml as unused.

This is causing a failure at system-definition configuration time.

Special case logic will need to be added to the connector pane
validation to ignore VIs provided by VeriStand.