Add explanation for possible deployment failure
The NI-Sync service will fail to start properly on Linux if EtherCAT
is selected as the primary time reference. This is something customers
would currently need to manually opt-in to, by following steps similar to
the ones outlined here: https://van.natinst.com/van/procedure/show/18231322.
While this may not be a typical use case, it was brought to my attention
by Alisha P. who spoke with Wendy Y. about it.

As a result of this service not starting properly, the first call we
make to niSync Initialize will fail. The action Alisha and I decided was
reasonable was to direct the user to view the log file which contains
more information. This commit conditionally adds this message to deployment
failures caused by the first niSync Initialize call.