Consolidate API to decrease OS-specific coupling (#7)
* Refactor API down to useful channels

Exposed channels provide a way to check the offset of the board time from the time reference, to check if a time reference is present, and to check if the system time is synchronized with the board time.

* Add Linux RT target

Conditionally disables unsupported driver calls on Linux. Disabling
these calls is done at the call site, without much regard for platform
abstraction. This can be cleaned up in the future. In areas where driver
support differs, devices other than the 6683 and 6683H have been made
into no-ops, although support for them remains in other areas of the
custom device.

The default compile behavior has been left as PharLap; it might be useful
to make the default case of platform-specific behavior error out
explicitly, to make it easier to add platforms in the future.

On PharLap the particular timing card synchronized with the system
clock is specified by name in a configuration file. On Linux, the
driver currently just uses the timing card in the lowest slot.

* Ignore errors from disconnecting terminals

On PharLap these never errored, so the error wire was passed through the
call. On Linux this will error for terminals which are invalid/not
connected. We want to disconnect unconditionally, so it is safe to
ignore the error.

* 1588 service is always running on Linux

We do not need to (and cannot) manually start 1588 on Linux in the same
way we do on PharLap. Setting the time reference is sufficient.

* Add Built/ directory to .gitignore

* Add initial support for 802.1AS

* Add explanation for possible deployment failure

The NI-Sync service will fail to start properly on Linux if EtherCAT
is selected as the primary time reference. This is something customers
would currently need to manually opt-in to, by following steps similar to
the ones outlined here: https://van.natinst.com/van/procedure/show/18231322.
While this may not be a typical use case, it was brought to my attention
by Alisha P. who spoke with Wendy Y. about it.

As a result of this service not starting properly, the first call we
make to niSync Initialize will fail. The action Alisha and I decided was
reasonable was to direct the user to view the log file which contains
more information. This commit conditionally adds this message to deployment
failures caused by the first niSync Initialize call.

Following a discussion with SEs, the conclusion was that maintaining an
upgrade path in this instance was not necessary, and since renaming or
moving channels would break any external references to them that a hard
break in compatibility would be preferable to a partial migration.

* Add new target to build spec