Work around issue with DAQmx devices as the master hardware synchronization device (#53)
Work around an initialization ordering issue related to the VeriStand engine's usage of DAQmx.

If a DAQmx module is configured as the chassis master hardware synchronization device, a DAQmx timing source is created prior to the custom device engine being launched. If we connect a new source to the 10MHz clock input at that point, we can cause the PLL to lose phase-lock and the timing source to be torn down, resulting in the primary VeriStand timed loop erroring. This issue doesn't always occur, as the new clock is sometimes "close-enough".

By connecting the clock terminals in ActionVIOnDownload, the chassis lock should be disciplined properly by the time a DAQmx timing source is created. We also connect them again in the engine, in case the original configuration failed (e.g. due to missing remote support) or something else went wrong. Double connecting does not cause any issues.